rename file injecting hints
